README
You just have to search what you want on Amazon, and then copy the link.

Run Amazon.py, and then it will aks for a link, and you just paste and enter.

I have scraped hundreds of items on Amazon with this program. If it does not find the element, it will probably input zero as value instead of return error.

I may write a bs4 version later as it is faster.

I also plan to write a program such that everything scraped from Amazon can immediately be imported to database like MySQL, but it takes time.

This program is far from perfect. You are welcome to contribute.